+++ src/Application/Get/BaseGetEntities.php
@@ -18,10 +18,16 @@ readonly abstract class BaseGetEntities
             $data = $entity->getData();
 
             foreach ($data as $key => $value) {
+
                 // Validate if is a related entity
-                if (str_starts_with($key, "_")) {
+                if (str_starts_with($key, "_") and is_array($value)) {
                     $data[$key] = $this->parseData($value);
                 }
+
+                if (str_starts_with($key, "_") and $value instanceof EntityData) {
+                    $data[$key] = ["entity" => $value->name, "data" => $value->getData()];
+                }
+
             }
             $output[] = ["entity" => $entity->name, "data" => $data];
         }